+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [Unreleased]
 ### Added
-- New functions: `asMap`, `getMapEntries`, `asMultiSet`
+- New functions: `asMap`, `getMapEntries`, `asMultiSet`, `getStringContent`
 - Support entries extraction from Dexx HashMap
 - Support entries extraction from vlsi.CompactHashMap